+++ component/common/loki/client/client_writeto_test.go
@@ -68,7 +68,7 @@ func TestClientWriter_LogEntriesAreReconstructedAndForwardedCorrectly(t *testing
 					Line:      line,
 				},
 			},
-		})
+		}, 0)
 	}
 
 	require.Eventually(t, func() bool {
@@ -125,7 +125,7 @@ func TestClientWriter_LogEntriesWithoutMatchingSeriesAreIgnored(t *testing.T) {
 					Line:      line,
 				},
 			},
-		})
+		}, 0)
 	}
 
 	time.Sleep(time.Second * 2)
@@ -237,7 +237,7 @@ func startWriter(segmentNum, seriesToReset int, target *clientWriteTo, lines int
 					Line:      fmt.Sprintf("%d - %d - hellooo", segmentNum, i),
 				},
 			},
-		})
+		}, 0)
 		// add some jitter between writes
 		randomSleepMax(time.Millisecond * 1)
 	}