+++ CHANGELOG.md
@@ -44,6 +44,10 @@ Main (unreleased)
 
 - Update version of River to support raw strings in flow using a backtick. (@erikbaranowski)
 
+- Added support for python profiling to `pyroscope.ebpf` component. (@korniltsev)
+ 
+- Add queueing logs remote write client for `loki.write` when WAL is enabled. (@thepalbi)
+
 ### Bugfixes
 
 - Fixed an issue where `loki.process` validation for stage `metric.counter` was